% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMeans.R
\name{KMeans}
\alias{KMeans}
\title{K-Means Clustering}
\usage{
KMeans(X, n_clusters, max_iterations = 100)
}
\arguments{
\item{X}{numeric matrix of data, or an object that can be coerced to such a matrix}

\item{n_clusters}{number of clusters}

\item{max_iterations}{number of iterations}
}
\value{
list containing centers and labels
}
\description{
K-Means Clustering
}
\examples{
X <- iris[, 1:4]
result <- KMeans(X, n_clusters = 3)
centers <- result$centers
labels <- result$labels
}
