% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_cv.R
\name{kmeans_cv}
\alias{kmeans_cv}
\title{Cross-validation for Kmeans}
\usage{
kmeans_cv(X, k, n_clusters_values)
}
\arguments{
\item{X}{input matrix}

\item{k}{number of centroids}

\item{n_clusters_values}{user-supplied number of clusters sequence for
hyper-parameter tuning}
}
\value{
best number of clustering
}
\description{
Cross-validation for Kmeans
}
\examples{
X <- iris[, 1:4]
n_clusters_values <- seq(2, 6)
kmeans_cv(X, k = 5, n_clusters_values)
}
