% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn0.R
\name{knn0}
\alias{knn0}
\title{k-Nearest Neighbour Classification}
\usage{
knn0(train, test, y, k, func = euclidean_distance, p = NULL)
}
\arguments{
\item{train}{matrix or data frame of training set cases}

\item{test}{matrix or data frame of test set cases}

\item{y}{factor of true classifications of training set}

\item{k}{number of neighbours considerd.a specified number}

\item{func}{optional parameters for distance function}

\item{p}{Minkowski power}
}
\value{
a single prediction of y
}
\description{
k-Nearest Neighbour Classification
}
\examples{
train = iris[1:120,]
test = iris[121:nrow(iris),]
knn0(train, test, 'Species', k = 5)
}
