% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn0_cv.R
\name{knn0_cv}
\alias{knn0_cv}
\title{Cross-validation for KNN}
\usage{
knn0_cv(data, k_values, num_folds)
}
\arguments{
\item{data}{input matrix}

\item{k_values}{user-supplied k sequence for hyper-parameter tuning}

\item{num_folds}{number of folds}
}
\value{
a table of k and correspond mean accuracy
}
\description{
Cross-validation for KNN
}
\examples{
iris_cv <- knn0_cv(iris, k_values = c(3,5,7,9), num_folds = 5)
iris_cv
}
