% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_bayes_cv.R
\name{naive_bayes_cv}
\alias{naive_bayes_cv}
\title{Cross-validation for train_naive_bayes}
\usage{
naive_bayes_cv(data, class_column, k = 5)
}
\arguments{
\item{data}{a matrix or data frame of categorical and numeric variables.}

\item{class_column}{a specified column representing class labels}

\item{k}{number of folds - default is 5}
}
\value{
average Cross-Validation Accuracy
}
\description{
Cross-validation for train_naive_bayes
}
\examples{
naive_bayes_cv(iris, class_column = "Species")
}
