% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_naive_bayes.R
\name{predict_naive_bayes}
\alias{predict_naive_bayes}
\title{make predictions form a 'train_naive_bayes' object}
\usage{
predict_naive_bayes(model, new_data)
}
\arguments{
\item{model}{fitted 'train_naive_bayes' object}

\item{new_data}{matrix of new values for x at which predictions are to be made}
}
\value{
a data frame containing the predicted class labels
}
\description{
make predictions form a 'train_naive_bayes' object
}
\examples{
train = iris[1:120,]
model <- train_naive_bayes(train, class_column = "Species")
predict_naive_bayes(model, new_data = iris[1:50,])
}
