% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_fit.R
\name{ridge_fit}
\alias{ridge_fit}
\title{fit a linear regression model with ridge regularization}
\usage{
ridge_fit(x, y, lambda = NULL)
}
\arguments{
\item{x}{input matrix, of dimension nobs x nvars}

\item{y}{response variable}

\item{lambda}{regularization parameter. can be a single number or a sequence
of lambda values}
}
\value{
a object containing call, lambda, and beta
}
\description{
fit a linear regression model with ridge regularization
}
\examples{
x = matrix(rnorm(100 * 20), 100, 20)
y = x[,1] + 2*x[,5]+ 6*x[,10] +rnorm(100)
fit <- ridge_fit(x, y, lambda = c(0.01, 0.02))
fit$beta
}
