% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_predict.R
\name{ridge_predict}
\alias{ridge_predict}
\title{make predictions from a "ridge_fit" object}
\usage{
ridge_predict(model, newdata, lambda = NULL)
}
\arguments{
\item{model}{fitted "ridge_fit" model object}

\item{newdata}{matrix of new values for x at which predictions are to be made}

\item{lambda}{value(s) of the regularization parameter lambda at which
predictions are required.}
}
\value{
an array of predicted value
}
\description{
make predictions from a "ridge_fit" object
}
\examples{
x = matrix(rnorm(100 * 20), 100, 20)
y = x[,1] + 2*x[,5]+ 6*x[,10] +rnorm(100)
fit <- ridge_fit(x,y,lambda = 0.01)
ridge_predict(fit, newdata = x[1:10, ])
}
