% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_naive_bayes.R
\name{train_naive_bayes}
\alias{train_naive_bayes}
\title{Naive Bayes Classifier}
\usage{
train_naive_bayes(train_data, class_column)
}
\arguments{
\item{train_data}{a matrix or data frame of categorical and numeric variables}

\item{class_column}{a specified column representing class labels}
}
\value{
a list containing the unique classes, means, and variances
}
\description{
Naive Bayes Classifier
}
\examples{
train = iris[1:120,]
model <- train_naive_bayes(train, class_column = "Species")
}
